Add transparent texture
This is step 1 of adding support for transparent textures.

A transparent texture extends from lib gdx textures but maintains an optimized
cache to quickly answer whether a pixel on a texture is transparent or not.

It extends (as opposed as composition) because it makes it easier to integrate
with libgdx objects (the transparent texture can be passed as a texture); this
also adopts the libgdx class organization (for ex. image extends from widget,
etc.).

In upcoming commits:
* A transparent texture will be added as component to building. This will
  eventually be necessary to process input events.
* A system will listen to building component creation events and create
  corresponding textures components
* For now the texture system will, at every frame, compute the texture component
  of each building depending on its state. This is essentially shifting and
  lifting the render() logic. The render() logic will pull the texture component
  of each building instead of computing it on the fly.
* On the click screen processing events, the reverse mapping (pixel to tile) of
  the click will be computed. Knowning the maximum height of a building
  (the max height of each texture), hit testing will be performed for each tile
  which is a collision candidate. See below:

Assuming the position of click (marked with '.'), and the max building height
indicated by the vertical arrow, the tiles that need to be tested for potential
collision are 0, and if no hit is detected, 1, and if no hit is detected, 2 is
the clicked tile. This is because 0 might contain a tall building hiding 1 and
2, and 1 might contain a tall building hiding 2. Collision testing is performed
by checking the transparent mask at the clicked coordinate, transposed in the
corresponding texture coordinates.

    x
   / \
  /   \   <--- position of click (.)   ^
 x  2 .x                               |
/ \   / \                              |
   \ /   \                             | max building height
    x  1  X                            |
   / \   /                             |
  /   \ /                              v
 x  0  x