Use LibGdx world actor to tick, draw and process input
Previous to this change we configured LibGdx with a custom input processor at
the highest level and a custom rendering loop. This is like telling LibGdx: "we
want to completely customize input processing and rendering".

This change integrates with LibGdx at a higher level. We define a Stage, which
is the equivalent of a Unity Scene. The main program delegates rendering and
input processing to the stage. It allows multiple stage easily (e.g other stages
could be main menu or tech tree, just achieved by swapping the stage).

A Stage is composed of multiple Actors, which each allow an act (tick) method,
hit detection and friendly input processing. For example the input processing is
passed to the actor with the same rendering coordinates so it's not necessary to
consider screen vertical flip anymore. Actors are organized as a tree.

Additionally we can use higher level Actors, such as Image, which have the
concept of layout. This will enable to organize the screen easily: the
WorldActor is the main rendering section (it can be considered as an animated
Image), but other elements of the layout could be side menu and buttons.

Overall this delegates more work to LibGdx and separates responsibilities
better.